+++ test/test_native_hw3/test_hw3_handler.cpp
@@ -271,7 +271,7 @@ void test_hw3_mux1_sends_exactly_1()
 
 void test_hw3_filter_ids_count()
 {
-    TEST_ASSERT_EQUAL_UINT8(6, handler.filterIdCount());
+    TEST_ASSERT_EQUAL_UINT8(7, handler.filterIdCount());
 }
 
 void test_hw3_filter_ids_values()
@@ -279,10 +279,11 @@ void test_hw3_filter_ids_values()
     const uint32_t *ids = handler.filterIds();
     TEST_ASSERT_EQUAL_UINT32(280, ids[0]);
     TEST_ASSERT_EQUAL_UINT32(390, ids[1]);
-    TEST_ASSERT_EQUAL_UINT32(921, ids[2]);
-    TEST_ASSERT_EQUAL_UINT32(1016, ids[3]);
-    TEST_ASSERT_EQUAL_UINT32(1021, ids[4]);
-    TEST_ASSERT_EQUAL_UINT32(2047, ids[5]);
+    TEST_ASSERT_EQUAL_UINT32(599, ids[2]);
+    TEST_ASSERT_EQUAL_UINT32(921, ids[3]);
+    TEST_ASSERT_EQUAL_UINT32(1016, ids[4]);
+    TEST_ASSERT_EQUAL_UINT32(1021, ids[5]);
+    TEST_ASSERT_EQUAL_UINT32(2047, ids[6]);
 }
 
 int main()