+++ test/test_native_hw4/test_hw4_handler.cpp
@@ -430,7 +430,7 @@ void test_hw4_gear_drive_clears_parked()
 
 void test_hw4_filter_ids_count()
 {
-    TEST_ASSERT_EQUAL_UINT8(6, handler.filterIdCount());
+    TEST_ASSERT_EQUAL_UINT8(7, handler.filterIdCount());
 }
 
 void test_hw4_filter_ids_values()
@@ -438,15 +438,17 @@ void test_hw4_filter_ids_values()
     const uint32_t *ids = handler.filterIds();
     TEST_ASSERT_EQUAL_UINT32(280, ids[0]);
 #if defined(ISA_SPEED_CHIME_SUPPRESS) && !defined(ESP32_DASHBOARD)
-    TEST_ASSERT_EQUAL_UINT32(921, ids[1]);
-    TEST_ASSERT_EQUAL_UINT32(923, ids[2]);
+    TEST_ASSERT_EQUAL_UINT32(599, ids[1]);
+    TEST_ASSERT_EQUAL_UINT32(921, ids[2]);
+    TEST_ASSERT_EQUAL_UINT32(923, ids[3]);
 #else
     TEST_ASSERT_EQUAL_UINT32(390, ids[1]);
-    TEST_ASSERT_EQUAL_UINT32(923, ids[2]);
+    TEST_ASSERT_EQUAL_UINT32(599, ids[2]);
+    TEST_ASSERT_EQUAL_UINT32(923, ids[3]);
 #endif
-    TEST_ASSERT_EQUAL_UINT32(1016, ids[3]);
-    TEST_ASSERT_EQUAL_UINT32(1021, ids[4]);
-    TEST_ASSERT_EQUAL_UINT32(2047, ids[5]);
+    TEST_ASSERT_EQUAL_UINT32(1016, ids[4]);
+    TEST_ASSERT_EQUAL_UINT32(1021, ids[5]);
+    TEST_ASSERT_EQUAL_UINT32(2047, ids[6]);
 }
 
 int main()