+++ test/test_native_legacy/test_legacy_handler.cpp
@@ -166,7 +166,7 @@ void test_legacy_ignores_unrelated_can_id()
 
 void test_legacy_filter_ids_count()
 {
-    TEST_ASSERT_EQUAL_UINT8(5, handler.filterIdCount());
+    TEST_ASSERT_EQUAL_UINT8(7, handler.filterIdCount());
 }
 
 void test_legacy_filter_ids_values()
@@ -175,8 +175,10 @@ void test_legacy_filter_ids_values()
     TEST_ASSERT_EQUAL_UINT32(69, ids[0]);
     TEST_ASSERT_EQUAL_UINT32(280, ids[1]);
     TEST_ASSERT_EQUAL_UINT32(390, ids[2]);
-    TEST_ASSERT_EQUAL_UINT32(921, ids[3]);
-    TEST_ASSERT_EQUAL_UINT32(1006, ids[4]);
+    TEST_ASSERT_EQUAL_UINT32(599, ids[3]);
+    TEST_ASSERT_EQUAL_UINT32(921, ids[4]);
+    TEST_ASSERT_EQUAL_UINT32(1006, ids[5]);
+    TEST_ASSERT_EQUAL_UINT32(1016, ids[6]);
 }
 
 int main()